pyproject.toml
[project]
name = "ainews"
version = "0.1.0"
description = "Daily AI news aggregator with summarization and static site generation"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.10",
    "lxml>=5.2.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "sentence-transformers>=2.2.2",
    "openai>=1.12.0",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "python-dateutil>=2.9.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[tool.black]
line-length = 88
target-version = ['py314']

[tool.ruff]
target-version = "py314"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.isort]
known-first-party = ["ainews"]
